Implementation of an efficient sampler for the Poisson distribution
This sampler is efficient for the special case when we know that we will draw many values from a Poisson with mean 'beta+count'.
I.e some fixed 'beta' value plus low counts. This is typically the case in LDA (Latent Dirichlet Allocation).
When this is the case, we can pre-generate WalkerAlias tables for counts less than some value (say 100). This means that
draws from the Poisson for count values < 100 will be very fast, where the standard Poisson draw is quite expensive.
For draws when lambda is bigger than 100, the Poisson draw can be well approximated with a normal distribution and will also be fast.