Added functionality for loading instances directly from a directory.
Now if the config option ’dataset’ points to a directory, the instances will be loaded (recursively) from this directory rather than from a file.
Same as in standard MALLET, it will use the directory names as labels and the filenames as instance names.
The ‘file_regex’ option regulates which files that will be included from the directory.