Finish implementation of abort functionality
Sending a 'kill -2 <PID>' to the process will  abort the sampler but let it finish
gracefully writing whatever stats that are requested in the config. It will wait
max 5 mins before it is finally killed but finish before if it natrurally does so.

When running in eclipse the '--runningInEclipse' option must be added, if so
the shutdown hook will NOT be installed since this doesn't work in Eclipse