Implemented Inference only program
You can train a sampler and set the "save_sampler" option. You can then
use the same config with this Inference only version and a new dataset,
and this program will then perform inference on the new documents based
on the previously trained model. The results are saved to the
doc_topic_mean_filename and doc_topic_theta_filename